+++ proto/src/main/java/team2/proto/controller/authentication/AuthenticationController.java
@@ -4,17 +4,17 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.web.bind.annotation.*;
-import team2.proto.dto.authentication.SigninDTO;
-import team2.proto.dto.authentication.JwtAuthenticationResponse;
+import org.springframework.web.multipart.MultipartFile;
+import team2.proto.dto.authentication.*;
+import team2.proto.entity.AdminPage;
 import team2.proto.entity.RefreshToken;
-import team2.proto.dto.authentication.RefreshTokenRequestDTO;
-import team2.proto.dto.authentication.UserSignUpDTO;
 import team2.proto.entity.User;
-import team2.proto.service.authentication.AuthenticationService;
-import team2.proto.service.authentication.JwtService;
-import team2.proto.service.authentication.RefreshTokenService;
+import team2.proto.service.authentication.*;
+import team2.proto.service.storage.StorageService;
 
 @RestController
 @RequestMapping("/api/auth")
@@ -25,6 +25,10 @@ public class AuthenticationController {
     private final RefreshTokenService refreshTokenService;
     private final JwtService jwtService;
     private final UserDetailsService userDetailService;
+    private final AuthenticationManager authenticationManager;
+    private final UserService userService;
+    private final StorageService storageService;
+    private final SchoolAuthorizationService schoolAuthorizationService;
 
     @PostMapping("/signup")
     public ResponseEntity<Void> signup(@RequestBody UserSignUpDTO params) {
@@ -42,8 +46,6 @@ public class AuthenticationController {
         }
         return ResponseEntity.ok(response);
     }
-
-
 
     @GetMapping("/logout")
     public ResponseEntity<String> logout(HttpServletRequest request) {
@@ -65,4 +67,52 @@ public class AuthenticationController {
                             .refreshToken(refreshTokenRequestDTO.getRefreshToken()).build();
                 }).orElseThrow(() -> new RuntimeException("Refresh Token is not in DB..!!"));
     }
+
+    @PostMapping("/withdrawal")
+    public ResponseEntity<Void> SignOut(@RequestBody WithdrawRequestDTO param, HttpServletRequest request) {
+        System.out.print("DEBUG >>>> UserController::SignOut");
+        User user = jwtService.extractUserFromRequest(request);
+        if (user == null) {
+            System.out.println("현재 로그인 상태가 아닙니다");
+            return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
+        }
+
+        // 해당 유저/이메일 정보가 DB에 정상적으로 존재하는지 판별하는 spring security 함수
+        // 인증 실패시 예외를 발생시킵니다.
+        authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(user.getEmail(), param.getPassword()));
+
+        // 먼저 로그아웃 처리
+        refreshTokenService.logoutRefreshToken(user.getId());
+        // deleteYn -> false로 바꿔주는 service
+        userService.deleteUser(user);
+
+        return new ResponseEntity<Void>(HttpStatus.OK);
+    }
+
+    @PostMapping("/school_auth")
+    public ResponseEntity<Void> SchoolAuth(@RequestParam(name = "image") MultipartFile imageFile, HttpServletRequest request) {
+        System.out.println("DEBUG >>>> AuthenticationController::SchoolAuth");
+        User user = jwtService.extractUserFromRequest(request);
+
+        // /proto/auth-images 폴더에 이미지 파일 저장합니다.
+        String photoUrl = storageService.store(imageFile);
+
+        // AdminPage 객체를 생성합니다.
+        AdminPage adminPage = AdminPage.builder()
+                .photoUrl(photoUrl)
+                .user(user)
+                .build();
+
+        // 코드 가독성을 위해 repository를 직접 호출하는 것 대신 service 불러오기
+        // adminPageRepository.save(adminPage);
+        try {
+            schoolAuthorizationService.save(adminPage);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return null;
+    }
+
 }
