+++ proto/src/main/java/team2/proto/repository/admin/AdminPageRepository.java
@@ -1,4 +1,4 @@
-package team2.proto.repository.user;
+package team2.proto.repository.admin;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
