+++ proto/src/main/java/team2/proto/service/authentication/SchoolAuthorizationService.java
@@ -1,0 +1,7 @@
+package team2.proto.service.authentication;
+
+import team2.proto.entity.AdminPage;
+
+public interface SchoolAuthorizationService {
+    public void save(AdminPage adminPage) throws Exception;
+}
